# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/common/upload/single": {
      "post": {
        "operationId": "UploadController_singleFileUpload",
        "summary": "文件上传",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk/uploadId": {
      "get": {
        "operationId": "UploadController_getChunkUploadId",
        "summary": "获取切片上传任务Id",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk": {
      "post": {
        "operationId": "UploadController_chunkFileUpload",
        "summary": "文件切片上传",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk/merge": {
      "post": {
        "operationId": "UploadController_chunkMergeFile",
        "summary": "合并切片",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkMergeFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/chunk/result": {
      "get": {
        "operationId": "UploadController_getChunkUploadResult",
        "summary": "获取切片上传结果",
        "parameters": [
          {
            "name": "uploadId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/common/upload/cos/authorization": {
      "get": {
        "operationId": "UploadController_getAuthorization",
        "summary": "获取cos上传密钥",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "通用-文件上传"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
    "/system/user/profile": {
      "get": {
        "operationId": "UserController_profile",
        "summary": "个人中心-用户信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "UserController_updateProfile",
        "summary": "个人中心-修改用户信息",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/profile/avatar": {
      "post": {
        "operationId": "UserController_avatar",
        "summary": "个人中心-上传用户头像",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/profile/updatePwd": {
      "put": {
        "operationId": "UserController_updatePwd",
        "summary": "个人中心-修改密码",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePwdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user": {
      "post": {
        "operationId": "UserController_create",
        "summary": "用户-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "get": {
        "operationId": "UserController_findPostAndRoleAll",
        "summary": "用户-角色+岗位",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "UserController_update",
        "summary": "用户-更新",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/list": {
      "get": {
        "operationId": "UserController_findAll",
        "summary": "用户-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deptId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nickName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phonenumber",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/deptTree": {
      "get": {
        "operationId": "UserController_deptTree",
        "summary": "用户-部门树",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/authRole/{id}": {
      "get": {
        "operationId": "UserController_authRole",
        "summary": "用户-分配角色-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/authRole": {
      "put": {
        "operationId": "UserController_updateAuthRole",
        "summary": "用户-角色信息-更新",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/{userId}": {
      "get": {
        "operationId": "UserController_findOne",
        "summary": "用户-详情",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/changeStatus": {
      "put": {
        "operationId": "UserController_changeStatus",
        "summary": "用户-停用角色",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/resetPwd": {
      "put": {
        "operationId": "UserController_resetPwd",
        "summary": "用户-重置密码",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPwdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/{id}": {
      "delete": {
        "operationId": "UserController_remove",
        "summary": "用户-删除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/system/user/export": {
      "post": {
        "operationId": "UserController_export",
        "summary": "导出用户信息数据为xlsx",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://nest-admin.dooring.vip/prod-api"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "FileUploadDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "file"
        ]
      },
      "ChunkMergeFileDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          }
        },
        "required": [
          "uploadId",
          "fileName"
        ]
      },
...
      "DateParamsDTO": {
        "type": "object",
        "properties": {}
      },
...
      "ChangeStatusDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "status"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "nickName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "sex": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "required": [
          "nickName",
          "email",
          "phonenumber",
          "sex"
        ]
      },
      "UpdatePwdDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "deptId": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "nickName": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "postIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "sex": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "postSort": {
            "type": "number"
          }
        },
        "required": [
          "nickName",
          "userName",
          "password",
          "postSort"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "deptId": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "nickName": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "postIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "sex": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "postSort": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          }
        },
        "required": [
          "userId"
        ]
      },
      "ResetPwdDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "password"
        ]
      },
      "ListUserDto": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "number",
            "description": "当前分页",
            "default": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页数量",
            "default": 10
          },
          "params": {
            "description": "时间范围",
            "allOf": [
              {
                "$ref": "#/components/schemas/DateParamsDTO"
              }
            ]
          },
          "orderByColumn": {
            "type": "string",
            "description": "排序字段"
          },
          "isAsc": {
            "type": "string",
            "description": "排序规则"
          },
          "deptId": {
            "type": "string"
          },
          "nickName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "pageNum",
          "pageSize"
        ]
      },
...
    }
  }
}